allow disabling hc per endpoint
